create task project and company models

+++ timer-billing-api/db/migrate/20170723090233_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[5.1]
       t.string :name
       t.string :email
       t.string :company_name
-      t.integer :bill_rate
+      t.float :bill_rate
       t.integer :auth_id
 
       t.timestamps
